refactor: Break monolithic tasks.py into modular tasks/ package
Splits the 2,377-line tasks.py into a well-organized package structure:

- tasks/core.py: TaskStatus enum, TASK_METADATA, TaskStatusManager, task_runner decorator
- tasks/config.py: Task configuration and history management functions
- tasks/fetch.py: Trip fetching tasks (periodic, manual, fetch_all_missing)
- tasks/coverage.py: Coverage update task
- tasks/maintenance.py: Cleanup, validation, and remapping tasks
- tasks/scheduler.py: Task scheduler logic
- tasks/management.py: Task management API functions
- tasks/webhook.py: Webhook processing task
- tasks/routes.py: Optimal route generation task
- tasks/__init__.py: Re-exports all public APIs for backwards compatibility

Each module is now focused on a single responsibility, making the codebase
more maintainable and easier to navigate. All existing imports continue
to work via the __init__.py re-exports.
