+++ tasks/management.py
@@ -1,0 +1,563 @@
+"""Task management API functions.
+
+This module provides functions for managing tasks through the API:
+- get_all_task_metadata: Retrieves metadata for all tasks with current status
+- manual_run_task: Manually triggers a task
+- trigger_manual_fetch_trips_range: Triggers a manual trip fetch for a date range
+- force_reset_task: Forcefully resets a stuck task
+- update_task_schedule: Updates task scheduling configuration
+- trigger_fetch_all_missing_trips: Triggers the fetch all missing trips task
+"""
+
+from __future__ import annotations
+
+import asyncio
+import uuid
+from datetime import UTC, datetime, timedelta
+from typing import Any
+
+from celery.utils.log import get_task_logger
+
+from celery_app import app as celery_app
+from date_utils import ensure_utc, parse_timestamp
+from db import (
+    serialize_datetime,
+    task_config_collection,
+    task_history_collection,
+    update_one_with_retry,
+)
+from tasks.config import check_dependencies, get_task_config, update_task_history_entry
+from tasks.core import TASK_METADATA, TaskStatus, TaskStatusManager
+from tasks.fetch import fetch_all_missing_trips
+
+logger = get_task_logger(__name__)
+
+
+async def get_all_task_metadata() -> dict[str, Any]:
+    """Retrieves metadata for all defined tasks, enriched with current status and configuration.
+
+    Returns:
+        A dictionary where keys are task IDs and values are dictionaries
+        containing task metadata and current status information.
+    """
+    try:
+        task_config = await get_task_config()
+        task_metadata_with_status = {}
+
+        for task_id, metadata in TASK_METADATA.items():
+            task_entry = metadata.copy()
+
+            config_data = task_config.get("tasks", {}).get(task_id, {})
+
+            estimated_next_run = None
+            last_run_any = config_data.get("last_run")
+            interval_minutes = config_data.get(
+                "interval_minutes",
+                metadata.get("default_interval_minutes"),
+            )
+            last_run = parse_timestamp(last_run_any)
+
+            if last_run and interval_minutes and interval_minutes > 0:
+                estimated_next_run = last_run + timedelta(minutes=interval_minutes)
+
+            task_entry.update(
+                {
+                    "enabled": config_data.get("enabled", True),
+                    "interval_minutes": interval_minutes,
+                    "status": config_data.get("status", TaskStatus.IDLE.value),
+                    "last_run": serialize_datetime(last_run),
+                    "next_run": serialize_datetime(estimated_next_run),
+                    "last_error": config_data.get("last_error"),
+                    "start_time": serialize_datetime(
+                        config_data.get("start_time"),
+                    ),
+                    "end_time": serialize_datetime(
+                        config_data.get("end_time"),
+                    ),
+                    "last_updated": serialize_datetime(
+                        config_data.get("last_updated"),
+                    ),
+                    "manual_only": metadata.get("manual_only", False),
+                },
+            )
+            task_metadata_with_status[task_id] = task_entry
+
+        return task_metadata_with_status
+    except Exception as e:
+        logger.exception("Error getting all task metadata: %s", e)
+        fallback_metadata = {}
+        for task_id, metadata in TASK_METADATA.items():
+            fallback_metadata[task_id] = {
+                **metadata,
+                "enabled": True,
+                "interval_minutes": metadata.get("default_interval_minutes"),
+                "status": TaskStatus.IDLE.value,
+                "last_run": None,
+                "next_run": None,
+                "last_error": "Error fetching status",
+                "start_time": None,
+                "end_time": None,
+                "last_updated": None,
+                "manual_only": metadata.get("manual_only", False),
+            }
+        return fallback_metadata
+
+
+async def manual_run_task(task_id: str) -> dict[str, Any]:
+    """Manually triggers one or all specified background tasks via Celery.
+
+    Args:
+        task_id: The identifier of the task to run, or 'ALL' to run all enabled tasks.
+
+    Returns:
+        A dictionary indicating the status of the trigger operation.
+    """
+    task_mapping = {
+        "periodic_fetch_trips": "tasks.periodic_fetch_trips",
+        "cleanup_stale_trips": "tasks.cleanup_stale_trips",
+        "validate_trips": "tasks.validate_trips",
+        "remap_unmatched_trips": "tasks.remap_unmatched_trips",
+        "update_coverage_for_new_trips": "tasks.update_coverage_for_new_trips",
+    }
+
+    if task_id == "ALL":
+        config = await get_task_config()
+        enabled_tasks = [
+            t_name
+            for t_name, t_config in config.get("tasks", {}).items()
+            if t_config.get("enabled", True) and t_name in task_mapping
+        ]
+        logger.info("Manual run requested for ALL enabled tasks: %s", enabled_tasks)
+        results = []
+        for task_name in enabled_tasks:
+            single_result = await _send_manual_task(task_name, task_mapping[task_name])
+            results.append(single_result)
+            await asyncio.sleep(0.1)
+
+        success = all(r.get("success", False) for r in results)
+        return {
+            "status": ("success" if success else "partial_error"),
+            "message": f"Triggered {len(results)} tasks.",
+            "results": results,
+        }
+
+    if task_id in task_mapping:
+        logger.info("Manual run requested for task: %s", task_id)
+        result = await _send_manual_task(task_id, task_mapping[task_id])
+        return {
+            "status": ("success" if result.get("success") else "error"),
+            "message": result.get("message", f"Failed to schedule task {task_id}"),
+            "task_id": result.get("task_id"),
+        }
+
+    logger.error("Manual run requested for unknown task: %s", task_id)
+    return {
+        "status": "error",
+        "message": f"Unknown or non-runnable task ID: {task_id}",
+    }
+
+
+async def _send_manual_task(
+    task_name: str,
+    celery_task_string_name: str,
+) -> dict[str, Any]:
+    """Internal helper to check dependencies and send a single manual task to Celery.
+
+    Args:
+        task_name: The application-specific task name.
+        celery_task_string_name: The string name used to register the task with Celery.
+
+    Returns:
+        A dictionary indicating success/failure and the Celery task ID if successful.
+    """
+    status_manager = TaskStatusManager.get_instance()
+    try:
+        dependency_check = await check_dependencies(task_name)
+        if not dependency_check["can_run"]:
+            reason = dependency_check.get("reason", "Dependencies not met")
+            logger.warning("Manual run for %s skipped: %s", task_name, reason)
+            return {
+                "task": task_name,
+                "success": False,
+                "message": reason,
+            }
+
+        celery_task_id = f"{task_name}_manual_{uuid.uuid4()}"
+
+        result = celery_app.send_task(
+            celery_task_string_name,
+            task_id=celery_task_id,
+            queue="default",
+            kwargs={"manual_run": True},  # Pass manual_run flag
+        )
+
+        await status_manager.update_status(task_name, TaskStatus.PENDING.value)
+        await update_task_history_entry(
+            celery_task_id=celery_task_id,
+            task_name=task_name,
+            status=TaskStatus.PENDING.value,
+            manual_run=True,
+            start_time=datetime.now(UTC),
+        )
+
+        logger.info(
+            "Manually triggered task '%s' -> Celery task '%s' (ID: %s)",
+            task_name,
+            celery_task_string_name,
+            celery_task_id,
+        )
+        return {
+            "task": task_name,
+            "success": True,
+            "message": f"Task {task_name} scheduled successfully.",
+            "task_id": result.id,
+        }
+    except Exception as e:
+        logger.exception("Error sending manual task %s", task_name)
+        await status_manager.update_status(
+            task_name,
+            TaskStatus.FAILED.value,
+            error=f"Manual trigger failed: {e}",
+        )
+        return {
+            "task": task_name,
+            "success": False,
+            "message": str(e),
+        }
+
+
+async def trigger_manual_fetch_trips_range(
+    start_date: datetime,
+    end_date: datetime,
+    *,
+    map_match: bool,
+) -> dict[str, Any]:
+    """Schedule a manual trip fetch for a specific date range via Celery."""
+
+    status_manager = TaskStatusManager.get_instance()
+
+    start_utc = ensure_utc(start_date)
+    end_utc = ensure_utc(end_date)
+
+    if end_utc <= start_utc:
+        raise ValueError("End date must be after start date")
+
+    celery_task_id = f"manual_fetch_trips_range_{uuid.uuid4()}"
+    kwargs = {
+        "start_iso": start_utc.isoformat(),
+        "end_iso": end_utc.isoformat(),
+        "map_match": bool(map_match),
+        "manual_run": True,
+    }
+    try:
+        logger.info(
+            "Scheduling manual fetch task via Celery: ID=%s, Start=%s, End=%s, MapMatch=%s",
+            celery_task_id,
+            start_utc.isoformat(),
+            end_utc.isoformat(),
+            map_match,
+        )
+
+        result = celery_app.send_task(
+            "tasks.manual_fetch_trips_range",
+            task_id=celery_task_id,
+            queue="default",
+            kwargs=kwargs,
+        )
+
+        logger.info(
+            "Successfully sent manual fetch task to Celery. Result ID: %s",
+            result.id,
+        )
+
+        await status_manager.update_status(
+            "manual_fetch_trips_range",
+            TaskStatus.PENDING.value,
+        )
+
+        await update_task_history_entry(
+            celery_task_id=celery_task_id,
+            task_name="manual_fetch_trips_range",
+            status=TaskStatus.PENDING.value,
+            manual_run=True,
+            start_time=datetime.now(UTC),
+            result={
+                "start_date": start_utc.isoformat(),
+                "end_date": end_utc.isoformat(),
+                "map_match": bool(map_match),
+            },
+        )
+
+        return {
+            "status": "success",
+            "message": f"Manual fetch scheduled (Task ID: {result.id})",
+            "task_id": result.id,
+        }
+    except Exception as exc:  # pragma: no cover - defensive scheduling
+        logger.exception("Failed to schedule manual fetch: %s", exc)
+        await status_manager.update_status(
+            "manual_fetch_trips_range",
+            TaskStatus.FAILED.value,
+            error=str(exc),
+        )
+        raise
+
+
+async def force_reset_task(
+    task_id: str,
+    reason: str | None = None,
+) -> dict[str, Any]:
+    """Forcefully reset a task's status to IDLE and revoke running Celery tasks.
+
+    This function:
+    1. Identifies any 'RUNNING' or 'PENDING' instances of the task in history.
+    2. Issues a Celery revoke command (terminate=True) for each.
+    3. Updates the history entries to 'REVOKED'.
+    4. Resets the global task configuration to 'IDLE'.
+    """
+
+    if task_id not in TASK_METADATA:
+        raise ValueError(f"Unknown task_id: {task_id}")
+
+    now = datetime.now(UTC)
+    message = reason or "Task force-stopped by user"
+    revoked_count = 0
+
+    # 1. Find running/pending instances in history
+    cursor = task_history_collection.find(
+        {
+            "task_id": task_id,
+            "status": {"$in": [TaskStatus.RUNNING.value, TaskStatus.PENDING.value]},
+        }
+    )
+
+    async for doc in cursor:
+        celery_task_id = doc["_id"]  # The _id is the celery_task_id
+        logger.warning(
+            "Force stopping task %s (Celery ID: %s)",
+            task_id,
+            celery_task_id,
+        )
+
+        # 2. Revoke the Celery task
+        try:
+            celery_app.control.revoke(celery_task_id, terminate=True)
+            revoked_count += 1
+        except Exception as e:
+            logger.error(
+                "Failed to revoke Celery task %s: %s",
+                celery_task_id,
+                e,
+            )
+
+        # 3. Update history entry
+        await update_task_history_entry(
+            celery_task_id=celery_task_id,
+            task_name=task_id,
+            status="REVOKED",
+            manual_run=doc.get("manual_run", False),
+            error=f"Force stopped: {message}",
+            end_time=now,
+        )
+
+    # 4. Reset global config
+    update_fields = {
+        f"tasks.{task_id}.status": TaskStatus.IDLE.value,
+        f"tasks.{task_id}.last_error": message,
+        f"tasks.{task_id}.end_time": now,
+        f"tasks.{task_id}.last_updated": now,
+        f"tasks.{task_id}.start_time": None,
+    }
+
+    await update_one_with_retry(
+        task_config_collection,
+        {"_id": "global_background_task_config"},
+        {"$set": update_fields},
+        upsert=True,
+    )
+
+    # Add a history entry for the force stop action itself
+    history_id = f"{task_id}_force_stop_{uuid.uuid4()}"
+    await update_task_history_entry(
+        celery_task_id=history_id,
+        task_name=task_id,
+        status="FORCED_STOP_ACTION",
+        manual_run=True,
+        error=f"User initiated force stop. Revoked {revoked_count} active instances.",
+        start_time=now,
+        end_time=now,
+        runtime_ms=0,
+    )
+
+    logger.info(
+        "Task %s force-reset by user. Revoked %d instances.",
+        task_id,
+        revoked_count,
+    )
+
+    return {
+        "status": "success",
+        "message": (
+            f"Task {task_id} was force stopped. "
+            f"Revoked {revoked_count} active instances."
+        ),
+        "task_id": task_id,
+        "revoked_count": revoked_count,
+    }
+
+
+async def update_task_schedule(task_config_update: dict[str, Any]) -> dict[str, Any]:
+    """Updates the task scheduling configuration (enabled status, interval) in the database.
+
+    Args:
+        task_config_update: A dictionary containing the updates. Can include:
+            'globalDisable': Boolean to disable/enable all tasks.
+            'tasks': A dictionary where keys are task IDs and values are dictionaries
+                     with 'enabled' (bool) or 'interval_minutes' (int) settings.
+
+    Returns:
+        A dictionary indicating the status of the update operation.
+    """
+    try:
+        global_disable_update = task_config_update.get("globalDisable")
+        tasks_update = task_config_update.get("tasks", {})
+        changes = []
+        update_payload = {}
+
+        if global_disable_update is not None:
+            if isinstance(global_disable_update, bool):
+                update_payload["disabled"] = global_disable_update
+                changes.append(
+                    f"Global scheduling disable set to {global_disable_update}"
+                )
+            else:
+                logger.warning(
+                    "Ignoring non-boolean value for globalDisable: %s",
+                    global_disable_update,
+                )
+
+        if tasks_update:
+            current_config = await get_task_config()
+            current_tasks = current_config.get("tasks", {})
+
+            for task_id, settings in tasks_update.items():
+                if task_id in TASK_METADATA:
+                    current_settings = current_tasks.get(task_id, {})
+
+                    if "enabled" in settings:
+                        new_val = settings["enabled"]
+                        if isinstance(new_val, bool):
+                            old_val = current_settings.get("enabled", True)
+                            if new_val != old_val:
+                                update_payload[f"tasks.{task_id}.enabled"] = new_val
+                                changes.append(
+                                    f"Task '{task_id}' enabled status: "
+                                    f"{old_val} -> {new_val}",
+                                )
+                        else:
+                            logger.warning(
+                                "Ignoring non-boolean value for enabled "
+                                "status of task '%s': %s",
+                                task_id,
+                                new_val,
+                            )
+
+                    if "interval_minutes" in settings:
+                        try:
+                            new_val = int(settings["interval_minutes"])
+                            if new_val <= 0:
+                                logger.warning(
+                                    "Ignoring invalid interval <= 0 for task '%s': %s",
+                                    task_id,
+                                    new_val,
+                                )
+                                continue
+                        except (ValueError, TypeError):
+                            logger.warning(
+                                "Ignoring non-integer interval for task '%s': %s",
+                                task_id,
+                                settings["interval_minutes"],
+                            )
+                            continue
+
+                        old_val = current_settings.get(
+                            "interval_minutes",
+                            TASK_METADATA[task_id]["default_interval_minutes"],
+                        )
+                        if new_val != old_val:
+                            update_payload[f"tasks.{task_id}.interval_minutes"] = (
+                                new_val
+                            )
+                            changes.append(
+                                f"Task '{task_id}' interval: {old_val} -> {new_val} mins",
+                            )
+                else:
+                    logger.warning(
+                        "Attempted to update configuration for unknown task: %s",
+                        task_id,
+                    )
+
+        if not update_payload:
+            logger.info("No valid configuration changes detected in update request.")
+            return {
+                "status": "success",
+                "message": "No valid configuration changes detected.",
+            }
+
+        result = await update_one_with_retry(
+            task_config_collection,
+            {"_id": "global_background_task_config"},
+            {"$set": update_payload},
+            upsert=True,
+        )
+
+        if result.modified_count > 0 or result.upserted_id is not None:
+            log_msg = (
+                f"Task configuration updated: {'; '.join(changes)}"
+                if changes
+                else "Task configuration updated (specific changes not detailed)."
+            )
+            logger.info(log_msg)
+            return {
+                "status": "success",
+                "message": "Task configuration updated successfully.",
+                "changes": changes,
+            }
+
+        logger.info(
+            "Task configuration update requested, but no document was "
+            "modified (values might be the same).",
+        )
+        return {
+            "status": "success",
+            "message": (
+                "No changes applied to task configuration (values may already match)."
+            ),
+        }
+    except Exception as e:
+        logger.exception("Error updating task schedule: %s", e)
+        return {
+            "status": "error",
+            "message": f"Error updating task schedule: {e}",
+        }
+
+
+async def trigger_fetch_all_missing_trips(
+    start_date: str | None = None,
+) -> dict[str, Any]:
+    """Triggers the fetch_all_missing_trips task."""
+    task_id = "fetch_all_missing_trips"
+
+    # Check dependencies
+    dep_check = await check_dependencies(task_id)
+    if not dep_check["can_run"]:
+        raise ValueError(f"Cannot run task: {dep_check['reason']}")
+
+    # Trigger the task
+    task = fetch_all_missing_trips.delay(manual_run=True, start_iso=start_date)
+
+    return {
+        "status": "success",
+        "message": "Task triggered successfully",
+        "task_id": task.id,
+    }
